Add files via upload
This is the Swift bundle that contains all necessary code for the SHIFT Furniture iOS Application. To run, the user must have CocoaPods installed and must install the folder's podfile. Everything app-related is included in this bundle. Please note that it does not contain the separate Arduino code for syncing the mobile application with the SHIFT Pod hardware.