Add files via upload
This is the master Arduino file for controlling the SHIFT Pods, the hardware devices for my Stanford ME 310 team, A.I. Tek. This can be uploaded to the ESP32 development board onboard a pair of motorized SHIFT Pods to allow the SHIFT Furniture mobile app (App Store) to control the SHIFT Pods movements remotely via WiFi requests.