Add files via upload
This is the Arduino file that, when uploaded to the SHIFT Pod hardware (an ESP32 development module), commands the motorized SHIFT Pods' movement. This program takes in GET requests from the SHIFT Furniture mobile application over WiFi and translates them into physical motion of the SHIFT Pods. The SHIFT Furniture mobile application will be unable to control the SHIFT Pods without it.